requirements.txt
bcrypt==3.1.1
Flask==0.12
Flask-Admin==1.4.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.0
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
Markdown==2.6.8
peewee==2.8.5
python-slugify==1.2.1
wtf-peewee==0.2.6
WTForms==2.1
